Air gap resources qube + improve project installer
The reason a DispVM isn't used to download the content is because there
is no way to verify that the content downloaded from Github using git is
integral. For the ISO, we have the SHA-256, however, even that could be
tainted because it's being downloaded from the same source.

As a result, an attacker capable of infiltrating the DispVM with network
access (whether by Github being intercepted or serving malicious content
or a vulnerability in Git, cURL, the Linux TCP/IP stack or othersise)
could simply attach themselves onto the content we are downloading
before it's transferred to the resources qube and gain access to it as
well.

This makes it no more secure to use a DispVM rather than just
temporarily enabling network access for the resources qube initially.