Set network location to public due to Windows bug
There appears to be a bug where Windows sometimes does not set network
location to what is in the answer file. In these cases, the default
action will happen which is to open a dialog so the user can select one.
However, in the case a user does not select one (what would happen
during an automatic installation) it will default to "public" also known
as "other".

So, this at least keeps the network location consistent to being public
instead of it switching between public and home.

https://social.technet.microsoft.com/Forums/windows/en-US/c80273d7-0e89-4daf-9c95-f0ac78ce3861/unattendxml-network-location-setting-issue-windows-7?forum=w7itproinstall