Improve method of enabling test signing
This method waits until the BCD is at least one byte in size. Unlike the
previous method, this one waits until the BCD exists AND it has begun
being written to.

I hope for this to eliminate the (very) rare race condition in the
previous method where the installation could fail due to the "Windows
could not update the computer's boot configuration" error.