Don't disable Security Center service
Seems unnecessary now that we are using a better method for disabling
all Action Center notifications. Also, disabling Security Center didn't
completely block those notifications.

Having the service disabed just creates a bad UX because then if a user
vists the Action Center they will be greeted with a message telling them
to enable the Security Center service and if they click "Turn on now" it
will just throw an error because the service is disabled.