Set firewall to permissive instead of disabling it
Disabling the Windows Firewall service (MpsSvc) causes issues when doing
certain things through the GUI such as enabling RDP. This leads to a bad
UX.

Also, seeing as the Qubes firewall isn't as featureful and integrated
into Windows as the Windows Firewall it's not hard to see why somebody
might want to use the Windows Firewall. For example, if they want to
allow everything but block traffic coming from a specific application.

Therefore, it's better to just configure it to be permissive instead of
completely disabling it.

The BFE service is a dependency of Windows Firewall and so must be
enabled too.