Move disabling Windows Defender to optimize.bat
This is as a result of Windows 10 not supporting the option so to keep
it consistent we're going to disable Windows Defender in the optimize
post script for all versions of Windows.

If the DisableAntiSpyware entry is used in the Windows 10 answer file it
seems to work for a bit if you quickly go to look at the Windows
Defender settings in Windows Security, however, everything is quickly
turned on by Windows 10 after that.

https://docs.microsoft.com/en-us/windows-hardware/customize/desktop/unattend/security-malware-windows-defender-disableantispyware

Also, due to Tamper Protection being enabled by default in the latest
versions of Windows 10, we must do something hacky to disable Windows
Defender. Tamper Protection is a kernel-mode driver that filters out
certain actions which lead to the direct disabling of Windows Defender.

Also, note that this doesn't have any effect on Windows 7 because for at
least the ISO we're using, it never included an AV.