Move download-windows.sh to isos folder
We can now get rid of the cd into the isos directory making the script
less rigid. It also increases security because somebody who wants to
download new Window media is now more inclined to transfer the script to
a dispoable VM and use it there instead of opening up windows-mgmt to
the network.