Specify evaluation on some Win Server answer files
Despite the InstallFrom image name being the same, these can only be
used to fully automate Windows Server eval media because non-eval media
will require an entry that specifies a product key (even if we set it to
empty).

Windows Server 2008 R2 appears to be an exception because that media,
despite having the full non-eval images, never asks for a product key
and therefore does not require that entry in the answer file. A non-eval
version of Windows Server 2012 R2+ will ask for the product key though.