Add empty cpi:offlineImage tag
By default when creating a new answer file in Windows AIK/ADK (both
tested, for Windows 7/10) and save it directly after it includes this
tag, so for the sake of correctness we should too. (Of course, the user
opening the answer file in AIK/ADK still has to add the path to the
wim/clg file themselves (cpi:source attribute). (e.g. wim:<path_to_wim>
or catalog:<path_to_clg>)