Assure CEIP and WER is disabled
The following documentation claims the CEIPEnabled setting has no effect
on Server Core installations of Windows Server 2008 R2 (which is an OS
we support), however, I was not able to reproduce this. CEIPEnabled
works on that OS just fine. However, just to be safe I guess there is no
harm in disabling it for a second time in here too.

The way we have it disabled here is the same method of disabling CEIP
that the CEIPEnabled setting uses. And at that point, we might as well
assure WER is disabled too.

https://docs.microsoft.com/en-us/windows-hardware/customize/desktop/unattend/microsoft-windows-sqmapi-ceipenabled