Remove TLS 1.2 in use Chocolatey message
The Chocolatey installation script (hosted on chocolatey.org) does not
show this message on Windows 10, for example, where recent versions of
TLS is not an issue. Also, for OSs where it does show up such as Windows
7, it no longer exactly says TLS 1.2 is not in use but that it's unsure
and here are potential fixes. I think it use to say something more along
the lines of "TLS 1.2 is not in use" before Chocolatey disabled versions
of TLS prior to 1.2 on their website but not any longer.