Update README
I removed the "Limitations" part under "Security" because it was a little
unfair to say Windows is insecure just because it's not open source.

That opinion follows the "many eyes" theory whereby if there are many
people looking at the code, more bugs will be found. I still think this
is a very relevant point especially considering how popular Windows is
making it a prime target for adversaries.

However, there's also much more to security such as the abundence of
exploit mitigations Microsoft has introduced in Windows 10 dramatically
raising the bar for exploitation.

Besides that, all I'd say is Windows has a lot of exposed attack surface
by default which of course is not good for security.

I was going to add that on the "Limitations" section but at that point
it was just getting so long that I figured this wasn't the right place
to interject my thoughts of the security of Windows.

At the end of the day, it largely comes down to whether or not you trust
Microsoft. :)