Improve error handling
My main goal in trapping on EXIT was to hopefully make a ^C from the
qvm-create-windows-qube main script gracefully terminate the
create-media script. I assumed it wasn't exiting gracefully because
qvm-run didn't pass the ^C (130) exit code to create-media.sh causing
the INT trap not to be activated, however that is not the case. qvm-run
just seems to forcefully terminate whatever is running without giving it
a change to cleanup. An option for graceful exiting should perhaps be
added to qvm-run.

Nevertheless, I think this makes the code a bit less prone to breaking
so I will keep it.

Changing cleanup to clean_exit has also allowed me to add a general
failure message at which point it makes sense to move the general
success message there too.

Lastly, it's not a guarantee that just because an iso_device exists
means it will also be mounted (e.g. ^C) so I've accounted for that
possibility.