optimize.bat: fix disabling good services
Disabling the "lmhosts" service had the effect of breaking UNC paths. I
imagine disabling "SSDPSRV" breaks something useful too. In general,
disabling services shouldn't be done because they break Windows in weird
ways.

The only service we disable now is DiagTrack in spyless.bat which is
confirmed as only being used for telemetry so that's the only service we
disable now.