Create qvm-create-windows-qube
sudo is now added to both sed -i invocations.

Not feeling great on that from a security perspective...but since the single passed input of the VM name used in the sed statement was typed into dom0 by the *operator* anyway (who already has sudo access!), it leaves me just a tiny bit queasy. Probably best to copy the file out, change it and copy it back, then do the switcheroo, but then you have to maintain permissions, etc. So...recommendations welcome.