Fix correctness issue
An attacker-controlled filename passed to clean_file_timestamp could
control options of the "touch" command. Luckily, there's no real threat
here because all filenames are set by us and I don't see any option in
the "touch" command that has the potential for real harm. This isn't
necessary for the other commands including the "find" exec (at least on
the GNU implementation) but I'm just add -- to be safe.