Cleanup final shutdown method
I believe the original reason for doing it this way goes back a long
time to when we weren't air gapping Windows through out the installation
process. Basically, my concern was that Windows could download updates
and then try installing them during shutdown only for the shutdown timer
to go off and kill the VM if it took too long.

Also, the mentioned issue with packages installing vcredist shouldn't be
an issue now that we're cross-compiling QWT (VC is a version of the CRT
runtime only used in projects built with Visual Studio i.e. Visual C++
compiler).