Clean auto-qwt before putting in new installer
This caused a problem if you've never deleted this folder like me
because I still had the old Qubes R4.0 installer in there causing both
of them to be executed.

Also, it should generally be expected to provide a fresh state with each
unpack anyway.

Preferably this is atomic with renameat2 + RENAME_EXCHANGE or using
symlinks + deleting temp files whose modification date is ~5 minutes old
but there are also other places where something like that should be done
(e.g. post ISO) so I will leave it for now. You would also have to be
very unlucky to get that bad timing (it's never once happened to me but
it is possible if you run multiple qvm-create-windows-qube installations
at once).