Mido: Fix exit codes and overall error handling
Our previous exit code of 1 conflicted with some stuff (e.g. in argument
parsing).

Implement new exit handling trap that correctly exits with the exit code
of the last command OR exits with the signal sent to the shell process.
All POSIX compliant tested across Bash, ZSH, Dash, and Posh shells.

This also seems to fix the "Dash set -e bug". I'm not actually sure if
this was ever a Dash bug or if it's just a POSIX thing overall. Could
easily just be due to differences in how sha256sum responds to signals
over, for example, curl.

Improve curl error handling so signals sent to curl aren't treated as
"Server returned an error status". Also, detect command not found exit
statuses.

I was able to figure this all out because of this question I asked on
Stack Exchange:
https://unix.stackexchange.com/questions/752570/why-does-trap-passthough-zero-instead-of-the-signal-the-process-was-killed-wit