Fix accepting drives install
On Win11, AppActivate doesn't work reliably, sometimes it requires
manually clicking on the "Windows Security" window.
According to the documentation[1] (for VBA not VBS... but I haven't
found VBS's AppActivate docs), there is a second argument "wait" -
whether to wait for the calling application to get focus first.
Documentation says the default is False, but it behaves like it
actually is True. Set it explicitly to False (0), which fixes the issue.

[1] https://learn.microsoft.com/en-us/office/vba/Language/Reference/user-interface-help/appactivate-statement

QubesOS/qubes-issues#1861